[hotfix] rename the params package of rich model mapper from mode to mapper
